+++ Model.MVC/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Model MVC")]
+[assembly: AssemblyTitle("Platform Model MVC")]
 [assembly: AssemblyDescription("Platform Model MVC Pattern")]
 
 [assembly: Guid("5fd9830d-8c16-449a-92a4-adb0fcad4e4c")]