+++ Model.SPS.PlugIn/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Model SPS Plugin")]
+[assembly: AssemblyTitle("Platform Model SPS Plugin")]
 [assembly: AssemblyDescription("Platform Model SPS Plugin Pattern")]
 
 [assembly: Guid("ae3cebd2-a6eb-45f6-8024-e53330ac8bd4")]