+++ Model.SPS.Plugin.PCL/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyTitle("Model SPS Plugin (Portable)")]
+[assembly: AssemblyTitle("Platform Model SPS Plugin PCL")]
 [assembly: AssemblyDescription("Platform Model SPS Plugin Pattern (Portable)")]