+++ Model.SPS/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Model SPS")]
+[assembly: AssemblyTitle("Platform Model SPS")]
 [assembly: AssemblyDescription("Platform Model SPS Pattern")]
 
 [assembly: Guid("a2c87538-0a4a-48f7-8728-9897c0cc015f")]