+++ Model/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Model Core")]
+[assembly: AssemblyTitle("Platform Model Core")]
 [assembly: AssemblyDescription("Platform Model Pattern")]
 
 [assembly: Guid("fbb674ee-ff22-4052-9621-b67effba7320")]