+++ Presentation.Reports/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Presentation Reporting")]
+[assembly: AssemblyTitle("Platform Presentation Reporting")]
 [assembly: AssemblyDescription("Platform Presentation for Reporting")]
 
 [assembly: Guid("c56b9867-7dc4-45b2-b314-fa599480e7fc")]