+++ Support.Data/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Support Data")]
+[assembly: AssemblyTitle("Platform Support Data")]
 [assembly: AssemblyDescription("Platform Support for Data")]
 
 [assembly: Guid("ee0007cf-f37f-4abc-b438-858998d7e2e8")]