+++ Support.PCL/Properties/AssemblyInfo.cs
@@ -1,6 +1,4 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Platform Support Portable")]
+[assembly: AssemblyTitle("Platform Support PCL")]
 [assembly: AssemblyDescription("Platform Support Core (Portable)")]