+++ Support.Web/Properties/AssemblyInfo.cs
@@ -1,8 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Support Web")]
+[assembly: AssemblyTitle("Platform Support Web")]
 [assembly: AssemblyDescription("Platform Support for Web")]
 
 [assembly: Guid("1bd4483f-f368-42cb-b6a6-ba58ba0c2583")]