+++ Support.Windows/Properties/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Support Windows")]
+[assembly: AssemblyTitle("Platform Support Windows")]
 [assembly: AssemblyDescription("Platform Support for Windows")]
 
 [assembly: Guid("42a71f9a-80c2-4572-89ab-30ee830d905a")]