+++ Support/Properties/AssemblyInfo.cs
@@ -1,10 +1,7 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("Support Core")]
+[assembly: AssemblyTitle("Platform Support Core")]
 [assembly: AssemblyDescription("Platform Support Core")]
 
 [assembly: Guid("5e416e35-cae1-46ab-b1c2-81926016b0f1")]
-
-